feature: implementación de gestión de usuarios y empleados, configuración de seguridad personalizada y seeding de datos iniciales

+++ src/main/java/com/luxStock/demo/entity/Sede.java
@@ -14,7 +14,7 @@ public class Sede {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_sede")
-    private Long idSede;
+    private Integer idSede;
     @Column(nullable = false, length = 100)
     private String nombre;
